requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
flask-cors
marshmallow-sqlalchemy
python-dotenv
passlib
tox
celery[redis]>=5.0.0
gunicorn
boto3
Authlib==0.14.3
psycopg2-binary
Flask-Authorize